rtx: remove second gamma correction
On stream E180 after working on it for almost a year we found out that we're doing double gamma correction. Our swapchain colorspace is already SRGB, so no need to sqrt() in shaders.

Also, comment on the fact that we need to HDRize skyboxes, and leave the curren hack intact.