Experimental: UTF-8 multilingual font rendering using FreeType
It is already possible to implement drawing, but the width and height retrieval needs to be redesigned. The current version contains an unknown bug.

Note: Since I am not familiar with the source code, any path-related parts need to be modified:

You need to manually modify "engine/wscript[line:153]" to specify the folder where "freetype.lib" is stored.
You need to manually modify "cl_font.c[line:120]" to specify the path of the TTF font file.